all: Use connect(2) + replace socket fd strategy
This commit removes the server's SocketRequest API. Instead, the
client now passes its socket file desriptor to the daemon which
checks the address family and socket type. If the checks pass,
a new socket is created with the same settings and the daemon
calls connect(2) on it. The new socket fd is then sent back
to the client, which replaces the original fd using dup2(2).

The ConnectRequest object has been refactored to hold an IPv4 or
IPv6 address and related data (I only just learned of the horrors
of the socketaddr types and casting between them).

With these changes, we can now implement configurable policy
to allow or deny connections based on the target address.